requirements.txt
streamlit==1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
Pillow>=10.0.0
streamlit-drawable-canvas==0.9.2
scipy>=1.11.0